[nemo-qml-plugin-dbus] Fix property setting/getting. Contributes to JB#20418
In some instances the property variant may contain a QJSValue instead
of a direct value. Detect this case and convert the QJSValue into a
variant.

The QDBusArgument::beginMap() overload without parameters is only used
for reading. Use the writing overload when generating the return value
of org.freedesktop.DBus.Properties.GetAll.

Don't attempt to add invalid variants to the property map as DBus does
not support this.

Add test for testing GetAll.